VectorDataWidget : Expose `_TableView` for other internal GafferUI uses
This just moves the class to its own file, no actual code has changed, except the stylesheet updates to match the new class identity.
